Update Mapbox branch

+++ qgis2web/utils.py
@@ -21,8 +21,8 @@ import re
 import shutil
 import sys
 from io import StringIO
-from PyQt5.QtCore import QDir, QVariant
-from PyQt5.QtGui import QPainter
+from qgis.PyQt.QtCore import QDir, QVariant
+from qgis.PyQt.QtGui import QPainter
 from qgis.core import (QgsApplication,
                        QgsProject,
                        QgsCoordinateReferenceSystem,
@@ -38,11 +38,14 @@ from qgis.core import (QgsApplication,
                        QgsCategorizedSymbolRenderer,
                        QgsGraduatedSymbolRenderer,
                        QgsRuleBasedRenderer,
+                       QgsNullSymbolRenderer,
                        QgsVectorFileWriter,
                        QgsRasterFileWriter,
                        QgsRasterPipe,
-                       QgsWkbTypes)
-from qgis.core import (Qgs25DRenderer, QgsGeometryGeneratorSymbolLayer)
+                       QgsMessageLog,
+                       QgsWkbTypes,
+                       Qgs25DRenderer,
+                       QgsGeometryGeneratorSymbolLayer)
 from qgis.utils import Qgis
 import processing
 import tempfile
@@ -53,14 +56,46 @@ ALL_ATTRIBUTES = 1
 TYPE_MAP = {
     QgsWkbTypes.Point: 'Point',
     QgsWkbTypes.Point25D: 'Point',
+    QgsWkbTypes.PointZ: 'Point',
+    QgsWkbTypes.PointM: 'Point',
+    QgsWkbTypes.PointZM: 'Point',
     QgsWkbTypes.LineString: 'LineString',
     QgsWkbTypes.LineStringM: 'LineString',
+    QgsWkbTypes.LineStringZ: 'LineString',
+    QgsWkbTypes.LineStringZM: 'LineString',
     QgsWkbTypes.LineString25D: 'LineString',
+    QgsWkbTypes.CircularString: 'LineString',
+    QgsWkbTypes.CircularStringZ: 'LineString',
+    QgsWkbTypes.CircularStringM: 'LineString',
+    QgsWkbTypes.CircularStringZM: 'LineString',
+    QgsWkbTypes.CompoundCurveZ: 'LineString',
+    QgsWkbTypes.CompoundCurveM: 'LineString',
+    QgsWkbTypes.CompoundCurveZM: 'LineString',
+    QgsWkbTypes.MultiLineStringZ: 'LineString',
+    QgsWkbTypes.MultiLineStringM: 'LineString',
+    QgsWkbTypes.MultiCurve: 'LineString',
+    QgsWkbTypes.MultiCurveM: 'LineString',
+    QgsWkbTypes.MultiCurveZ: 'LineString',
+    QgsWkbTypes.MultiCurveZM: 'LineString',
     QgsWkbTypes.Polygon: 'Polygon',
+    QgsWkbTypes.PolygonZ: 'Polygon',
+    QgsWkbTypes.PolygonM: 'Polygon',
     QgsWkbTypes.PolygonZM: 'Polygon',
     QgsWkbTypes.Polygon25D: 'Polygon',
+    QgsWkbTypes.CurvePolygon: 'Polygon',
+    QgsWkbTypes.CurvePolygonZ: 'Polygon',
+    QgsWkbTypes.CurvePolygonM: 'Polygon',
+    QgsWkbTypes.CurvePolygonZM: 'Polygon',
+    QgsWkbTypes.MultiPolygonZ: 'Polygon',
+    QgsWkbTypes.Triangle: 'Polygon',
+    QgsWkbTypes.TriangleZ: 'Polygon',
+    QgsWkbTypes.TriangleM: 'Polygon',
+    QgsWkbTypes.TriangleZM: 'Polygon',
     QgsWkbTypes.MultiPoint: 'MultiPoint',
     QgsWkbTypes.MultiPoint25D: 'MultiPoint',
+    QgsWkbTypes.MultiPointZ: 'MultiPoint',
+    QgsWkbTypes.MultiPointM: 'MultiPoint',
+    QgsWkbTypes.MultiPointZM: 'MultiPoint',
     QgsWkbTypes.MultiLineString: 'MultiLineString',
     QgsWkbTypes.MultiLineStringM: 'MultiLineString',
     QgsWkbTypes.MultiLineStringZ: 'MultiLineString',
@@ -68,7 +103,8 @@ TYPE_MAP = {
     QgsWkbTypes.MultiLineString25D: 'MultiLineString',
     QgsWkbTypes.MultiPolygon: 'MultiPolygon',
     QgsWkbTypes.MultiPolygon25D: 'MultiPolygon',
-    QgsWkbTypes.MultiPolygonZM: 'MultiPolygon'}
+    QgsWkbTypes.MultiPolygonZM: 'MultiPolygon',
+    QgsWkbTypes.MultiPolygonM: 'MultiPolygon'}
 
 MB_TYPE_MAP = {
     'Point': 'circle',
@@ -93,37 +129,39 @@ PLACEMENT = ['bottomleft', 'topleft', 'topright', 'bottomleft', 'bottomright']
 
 
 def tempFolder():
-    tempDir = os.path.join(unicode(QDir.tempPath()), 'qgis2web')
+    tempDir = os.path.join(QDir.tempPath(), 'qgis2web')
     if not QDir(tempDir).exists():
         QDir().mkpath(tempDir)
 
-    return unicode(os.path.abspath(tempDir))
+    return os.path.abspath(tempDir)
 
 
 def getUsedFields(layer):
     fields = []
     try:
         fields.append(layer.renderer().classAttribute())
-    except:
+    except Exception:
         pass
-    labelsEnabled = unicode(
-        layer.customProperty("labeling/enabled")).lower() == "true"
+    labelsEnabled = layer.customProperty("labeling/enabled").lower() == "true"
     if labelsEnabled:
         fields.append(layer.customProperty("labeling/fieldName"))
     return fields
 
 
 def writeTmpLayer(layer, restrictToExtent, iface, extent):
+    if layer.wkbType() == QgsWkbTypes.NoGeometry:
+        return
+
     fields = layer.fields()
     usedFields = []
     for count, field in enumerate(fields):
-        fieldIndex = fields.indexFromName(unicode(field.name()))
+        fieldIndex = fields.indexFromName(field.name())
         editorWidget = layer.editorWidgetSetup(fieldIndex).type()
         addField = False
         try:
             if layer.renderer().classAttribute() == field.name():
                 addField = True
-        except:
+        except Exception:
             pass
         if layer.customProperty("labeling/fieldName") == field.name():
             addField = True
@@ -136,16 +174,18 @@ def writeTmpLayer(layer, restrictToExtent, iface, extent):
     if crs.isValid():
         uri += '?crs=' + crs.authid()
     for field in usedFields:
-        fieldIndex = layer.fields().indexFromName(unicode(
-            layer.fields().field(field).name()))
+        fieldIndex = layer.fields().indexFromName(
+            layer.fields().field(field).name())
         editorWidget = layer.editorWidgetSetup(fieldIndex).type()
         fieldType = layer.fields().field(field).type()
         fieldName = layer.fields().field(field).name()
         if (editorWidget == 'Hidden'):
             fieldName = "q2wHide_" + fieldName
-        fieldType = "double" if (fieldType == QVariant.Double or
-                                 fieldType == QVariant.Int) else ("string")
-        uri += '&field=' + unicode(fieldName) + ":" + fieldType
+        if fieldType == QVariant.Double or fieldType == QVariant.Int:
+            fieldType = "double"
+        else:
+            fieldType = "string"
+        uri += '&field=' + fieldName + ":" + fieldType
     newlayer = QgsVectorLayer(uri, layer.name(), 'memory')
     writer = newlayer.dataProvider()
     outFeat = QgsFeature()
@@ -157,7 +197,7 @@ def writeTmpLayer(layer, restrictToExtent, iface, extent):
         try:
             transform = QgsCoordinateTransform(canvasCRS, layerCRS,
                                                QgsProject.instance())
-        except:
+        except Exception:
             transform = QgsCoordinateTransform(canvasCRS, layerCRS)
         projectedExtent = transform.transformBoundingBox(extent)
         request = QgsFeatureRequest(projectedExtent)
@@ -182,7 +222,7 @@ def exportLayers(iface, layers, folder, precision, optimize, popupField, json,
     QDir().mkpath(layersFolder)
     for count, (layer, encode2json, popup) in enumerate(zip(layers, json,
                                                             popupField)):
-        sln = safeName(layer.name()) + "_" + unicode(count)
+        sln = safeName(layer.name()) + "_" + str(count)
         vts = layer.customProperty("VectorTilesReader/vector_tile_source")
         if (layer.type() == layer.VectorLayer and vts is None and
                 (layer.providerType() != "WFS" or encode2json)):
@@ -209,19 +249,27 @@ def exportVector(layer, sln, layersFolder, restrictToExtent, iface,
     path = os.path.join(layersFolder, sln + ".js")
     options = []
     if precision != "maintain":
-        options.append("COORDINATE_PRECISION=" + unicode(precision))
-    QgsVectorFileWriter.writeAsVectorFormat(cleanLayer, tmpPath, "utf-8", crs,
-                                            'GeoJson', 0,
-                                            layerOptions=options)
-    with open(path, mode="w", encoding="utf8") as f:
-        f.write("var %s = " % ("json_" + sln))
-        with open(tmpPath, encoding="utf8") as tmpFile:
-            for line in tmpFile:
-                if minify:
-                    line = line.strip("\n\t ")
-                    line = removeSpaces(line)
-                f.write(line)
-    os.remove(tmpPath)
+        options.append("COORDINATE_PRECISION=" + str(precision))
+    e, err = QgsVectorFileWriter.writeAsVectorFormat(cleanLayer, tmpPath,
+                                                     "utf-8", crs, 'GeoJson',
+                                                     0, layerOptions=options)
+    if e == QgsVectorFileWriter.NoError:
+        with open(path, mode="w", encoding="utf8") as f:
+            f.write("var %s = " % ("json_" + sln))
+            with open(tmpPath, encoding="utf8") as tmpFile:
+                for line in tmpFile:
+                    if minify:
+                        line = line.strip("\n\t ")
+                        line = removeSpaces(line)
+                    f.write(line)
+        os.remove(tmpPath)
+    else:
+        QgsMessageLog.logMessage(
+            "Could not write json file {}: {}".format(tmpPath, err),
+            "qgis2web",
+            level=Qgis.Critical)
+        return
+
     fields = layer.fields()
     for field in fields:
         exportImages(layer, field.name(), layersFolder + "/tmp.tmp")
@@ -277,14 +325,13 @@ def add25dAttributes(cleanLayer, layer, canvas):
 
 def exportRaster(layer, count, layersFolder, feedback, iface, matchCRS):
     feedback.showFeedback("Exporting %s to PNG..." % layer.name())
-    name_ts = (safeName(layer.name()) + unicode(count) +
-               unicode(int(time.time())))
+    name_ts = safeName(layer.name()) + str(count) + str(int(time.time()))
 
     # We need to create a new file to export style
     piped_file = os.path.join(tempfile.gettempdir(), name_ts + '_piped.tif')
 
     piped_extent = layer.extent()
-    piped_width = layer.height()
+    # piped_width = layer.height()
     piped_height = layer.width()
     piped_crs = layer.crs()
     piped_renderer = layer.renderer()
@@ -299,8 +346,9 @@ def exportRaster(layer, count, layersFolder, feedback, iface, matchCRS):
     file_writer.writeRaster(pipe, piped_height, -1, piped_extent, piped_crs)
 
     # Export layer as PNG
-    out_raster = os.path.join(layersFolder, safeName(layer.name()) + "_" +
-                              unicode(count) + ".png")
+    out_raster = os.path.join(layersFolder,
+                              safeName(layer.name()) + "_" +
+                              str(count) + ".png")
 
     projectCRS = iface.mapCanvas().mapSettings().destinationCrs()
     if not (matchCRS and layer.crs() == projectCRS):
@@ -310,25 +358,25 @@ def exportRaster(layer, count, layersFolder, feedback, iface, matchCRS):
         try:
             xform = QgsCoordinateTransform(crsSrc, crsDest,
                                            QgsProject.instance())
-        except:
+        except Exception:
             xform = QgsCoordinateTransform(crsSrc, crsDest)
         extentRep = xform.transformBoundingBox(layer.extent())
 
-        extentRepNew = ','.join([unicode(extentRep.xMinimum()),
-                                 unicode(extentRep.xMaximum()),
-                                 unicode(extentRep.yMinimum()),
-                                 unicode(extentRep.yMaximum())])
+        extentRepNew = ','.join([str(extentRep.xMinimum()),
+                                 str(extentRep.xMaximum()),
+                                 str(extentRep.yMinimum()),
+                                 str(extentRep.yMaximum())])
 
         # Reproject in 3857
         piped_3857 = os.path.join(tempfile.gettempdir(),
                                   name_ts + '_piped_3857.tif')
-        qgis_version = Qgis.QGIS_VERSION
+        # qgis_version = Qgis.QGIS_VERSION
 
         old_stdout = sys.stdout
         sys.stdout = mystdout = StringIO()
         try:
             processing.algorithmHelp("gdal:warpreproject")
-        except:
+        except Exception:
             pass
         sys.stdout = old_stdout
 
@@ -364,12 +412,12 @@ def exportRaster(layer, count, layersFolder, feedback, iface, matchCRS):
                 try:
                     k = line.split(":")[0]
                     warpArgs[k] = params[k]
-                except:
+                except Exception:
                     pass
 
         try:
             processing.run("gdal:warpreproject", warpArgs)
-        except:
+        except Exception:
             shutil.copyfile(piped_file, piped_3857)
 
         try:
@@ -392,13 +440,13 @@ def exportRaster(layer, count, layersFolder, feedback, iface, matchCRS):
                                               "COPY_SUBDATASETS": False,
                                               "OPTIONS": "",
                                               "OUTPUT": out_raster})
-        except:
+        except Exception:
             shutil.copyfile(piped_3857, out_raster)
     else:
-        srcExtent = ','.join([unicode(piped_extent.xMinimum()),
-                              unicode(piped_extent.xMaximum()),
-                              unicode(piped_extent.yMinimum()),
-                              unicode(piped_extent.yMaximum())])
+        srcExtent = ','.join([str(piped_extent.xMinimum()),
+                              str(piped_extent.xMaximum()),
+                              str(piped_extent.yMinimum()),
+                              str(piped_extent.yMaximum())])
         processing.run("gdal:translate", {"INPUT": piped_file,
                                           "OUTSIZE": 100,
                                           "OUTSIZE_PERC": True,
@@ -429,6 +477,8 @@ def is25d(layer, canvas, restrictToExtent, extent):
     if vts is not None:
         return False
     renderer = layer.renderer()
+    if isinstance(renderer, QgsNullSymbolRenderer):
+        return False
     if isinstance(renderer, Qgs25DRenderer):
         return True
     symbols = []
@@ -547,7 +597,7 @@ def exportImages(layer, field, layerFileName):
 
     for feature in layer.getFeatures(fr):
         photo_file_name = feature.attribute(field)
-        if type(photo_file_name) is not unicode:
+        if type(photo_file_name) is not str:
             continue
 
         source_file_name = photo_file_name
@@ -562,7 +612,7 @@ def exportImages(layer, field, layerFileName):
 
         try:
             shutil.copyfile(source_file_name, photo_file_name)
-        except IOError as e:
+        except IOError:
             pass
 
 
@@ -579,4 +629,63 @@ def handleHiddenField(layer, field):
 def getRGBAColor(color, alpha):
     r, g, b, a = color.split(",")
     a = (float(a) / 255) * alpha
-    return "'rgba(%s)'" % ",".join([r, g, b, unicode(a)])
+    return "'rgba(%s)'" % ",".join([r, g, b, str(a)])
+
+
+def boilType(fieldType):
+    fType = None
+    if fieldType.lower() in ["boolean", "bool"]:
+        fType = "bool"
+    if fieldType.lower() in ["double", "real", "decimal", "numeric"]:
+        fType = "real"
+    if fieldType.lower() in ["integer", "integer64", "uint",
+                             "int", "longlong", "int4", "ulonglong"]:
+        fType = "int"
+    if fieldType.lower() in ["char", "string", "text", "varchar", "nchar",
+                             "nvarchar"]:
+        fType = "str"
+    if fieldType.lower() in ["date"]:
+        fType = "date"
+    if fieldType.lower() in ["datetime", "timestamp",
+                             "timestamp without time zone"]:
+        fType = "datetime"
+    if fieldType.lower() in ["time"]:
+        fType = "time"
+    return fType
+
+
+def returnFilterValues(layer_list, fieldName, fieldType):
+    if fieldType.lower() == "bool":
+        return {"name": fieldName, "type": fieldType,
+                "values": ["true", "false"]}
+    filterValues = []
+    for layer in layer_list:
+        if layer.type() == layer.VectorLayer:
+            fields = layer.fields()
+            for f in fields:
+                if boilType(f.typeName()) == fieldType:
+                    if f.name() == fieldName:
+                        iterator = layer.getFeatures()
+                        for feature in iterator:
+                            if feature[fieldName] is not None:
+                                filterValues.append(feature[fieldName])
+    if filterValues == []:
+        return
+    if fieldType == "str":
+        cleanFilterValues = list(dict.fromkeys(filterValues))
+        cleanFilterValues.sort()
+    if fieldType == "int":
+        cleanFilterValues = [min(filterValues) if min(filterValues) >= 0
+                             else 0,
+                             max(filterValues) if max(filterValues) >= 0
+                             else 0]
+        if cleanFilterValues[0] == cleanFilterValues[1]:
+            cleanFilterValues[1] = cleanFilterValues[0] + 1
+    if fieldType in ["date", "time", "real", "datetime"]:
+        cleanFilterValues = [min(filterValues),
+                             max(filterValues)]
+        if cleanFilterValues[0] == cleanFilterValues[1]:
+            if fieldType == "real":
+                add = 1 / 10 * (cleanFilterValues[1] - cleanFilterValues[0])
+                cleanFilterValues[1] = cleanFilterValues[0] + add
+    return {"name": fieldName, "type": fieldType, "values": cleanFilterValues}
